Changes to dsPIC code and Documentation
Development of the Primary and Secondary V4 dsPIC code. This version is
not yet stable.

+++ Code/PIC_Projects/dsPIC/PrimaryV4.X/currentsensor.c
@@ -1,4 +1,6 @@
-/* Configure the ADC module to read the analog output from the current sensor.
+/* Configure the ADC module for manually triggered sampling and automatically
+ * triggered conversion from a single ADC channel. The ADC outputs an unsigned
+ * integer.
  */
 #include <libpic30.h>
 #include <p33EP512MC806.h>
@@ -16,7 +18,10 @@ void initialize_ADC(void) {
     AD1CON1bits.SSRCG    = 0; // Use an automatic trigger
     AD1CON2bits.VCFG    = 0;  // Ensure AVDD and AVSS are used as Vref+ and Vref-
     AD1CON3bits.SAMC    = 15; //
-    AD1CON3bits.ADCS = 9;     // Tad = 117.6ns = 10*Tcy,  Tcy = 12.5ns
+    AD1CON3bits.ADCS = 9;     // Tad = (ADCS + 1)*Tcy
+                              // Tcy = 1 / 40MHz = 25 ns
+                              // Tad = 10*25ns = 250ns
+                              // Datasheet sets Tad minimum at 117.6ns
 
     // Initialize MUXA Input Selection
     AD1CHS0bits.CH0SA = 8;
@@ -24,8 +29,8 @@ void initialize_ADC(void) {
     AD1CON1bits.ADON        = 1; //Turn on the ADC converter
 }
 
-int ADC_Read(void) { //manual sampling and conversion function
+unsigned short ADC_Read(void) { //manual sampling and conversion function
     AD1CON1bits.SAMP = 1; //Start sampling, sampling is stopped after 1us
     while (!AD1CON1bits.DONE); //wait for sampling and conversion to finish
-    return ADC1BUF0;
+    return (unsigned short) ADC1BUF0;
 }
