+++ Code/PIC_Projects/dsPIC/PrimaryV4.X/currentsensor.h
@@ -2,7 +2,7 @@
 #define	CURRENTSENSOR_H
 
 void initialize_ADC(void);
-int ADC_Read(void);
+unsigned short ADC_Read(void);
 
 #endif	/* CURRENTSENSOR_H */
